task#311: fix c7 system-connector tests
